components/yolo-trainer/requirements.txt
torch==2.1.0
torchvision==0.16.0
ultralytics
google-cloud-storage